First commit for the basic cryptic crossword tool
The current version of a crossword tool that fetches a user-specified
number of  cryptic crosswords from the Guardian newspaper's website in a
blind-accessible format (no grid).

Based on the user-input, a fixed number of crosswords are fetched
retroactively starting from a particular date , going backwards.

Possible features to add are:
a) Ability to choose crosswords to fetch by serial number, date
b) Fetching crosswords in a printable file format
c) Adding ability to display the grid in a graphical grid layout using
Python plotting tools
d) Calculating a difficulty rating for crossword, and clues
e) Fetching solutions to crosswords along with clues, etc.